requirements.txt
# Orchestr'IA - Phase 1 Dependencies

# Interface
streamlit>=1.30.0

# AI Provider
openai>=1.12.0
tiktoken>=0.6.0

# PDF extraction (fallback chain)
pymupdf>=1.23.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Document processing
python-docx>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
openpyxl>=3.1.0
pandas>=2.1.0

# DOCX export
lxml>=5.0.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Hashing / deduplication
# (hashlib is stdlib)

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
